If no operator is specified, don't filter
Previously, "=" would be used as the operator instead. But if the user didn't specify an operator, it's safer to assume she doesn't want to filter at all.

This issue arose for me using with HTML `type="range"` input, which always submits a value (even if the user doesn't touch the input field--in which case the starting value is used).
So, without doing anything, SonataAdmin would insist on only showing content objects who had that value = 0, forcing me to manually set the filter to ">=" on each page load.