Fix not working filters on model list popup
When using `sonata_type_model_list`, default filters are not working inside the list popup due to an AJAX call. This function is catching that website click, so we must ensure that the click href contains something different than *empty* and *#* to make the AJAX request. `return` is needed so others JS can catch that event. For example: Bootstrap dropdown (filters, as seen in previous lines)