Fix how dates from `DateRangeType` are handled by `DateRangeFilter`
When the `DateRangeType` uses different values between "view_timezone" and "model_timezone"
options, the values that are processed by `AbstractDateFilter::filter()` are `\DateTime`
objects (start and end) holding the same timestamp than the view data, but represented
in the timezone configured for the model data (thus, the offset of these objects
are set to the model timezone and the date and time references are adjusted accordingly).
Previous to this commit, these date and time conversions were becoming imprecise
since the `\DateTime` objects used in the filter were being manipulated to set the
time part to "00:00:00" (in case of start) and to "23:59:59" (in case of end),
which produces wrong results given the offset between the view and the model data
is lost in these operations.
This commit removes the manipulation of the start time completely, and changes
the manipulation done for the end time, by adding "23:59:59" hours to the received
value, since both dates are expressed originally at "00:00:00" in the view timezone.