Remove test
This test is using a Sonata\DoctrineORMAdminBundle\Tests\Filter\QueryBuilder instead of a Doctrine one.
This test, which seems complicated, is not really testing something (or something useful), there are some expects about the implementation, nothing about the result of this method.
There is already another test on the computeNbResult which is doing better assertions.